% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNS.R
\name{BNS}
\alias{BNS}
\title{Barndorff‐Nielsen and Shephard Realized Variance}
\usage{
BNS(ticker, start_date, end_date, prediction_period, days_to_drop, lambda1, delta1, gamma1, phi1, testing)
}
\arguments{
\item{ticker}{Stock ticker}

\item{start_date}{Starting date of data to use}

\item{end_date}{Ending date of data to use}

\item{prediction_period}{How many days to predict using the model}

\item{days_to_drop}{Number of days of realized variances to drop when fitting the model}

\item{lambda}{The mean reverting variable}

\item{delta}{Influences the tail thickness of the disribution}

\item{gamma}{Influences the scale of the distribution}

\item{phi}{Volatility of the volatility.}

\item{testing}{Whether to use prediction_period days prior to end_date as the prediction period. If false, the prediction period takes place after end_date}
}
\value{
Realized variance according to the Barndorff‐Nielsen and Shephard model when the correct parameter values are used
}
\description{
Barndorff‐Nielsen and Shephard Realized Variance
}
