% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heston.R
\name{Heston}
\alias{Heston}
\title{Heston Realized Variance and Volitility}
\usage{
Heston(ticker, start_date, end_date, prediction_period, days_to_drop, kappa, theta, gamma, variance, testing)
}
\arguments{
\item{ticker}{Stock ticker}

\item{start_date}{Starting date of data to use}

\item{end_date}{Ending date of data to use}

\item{prediction_period}{How many days to predict using the model}

\item{days_to_drop}{Number of days of realized variances to drop when fitting the model}

\item{kappa}{Mean reverting parameter.}

\item{theta}{The long term average.}

\item{gamma}{The volatiltiy of the volatility.}

\item{variance}{Boolean where true predicts variance and false predicts volitility}

\item{testing}{Whether to use prediction_period days prior to end_date as the prediction period. If false, the prediction period takes place after end_date}
}
\value{
Realized variance or volitility according to the Heston model when the correct parameter values are used
}
\description{
Heston Realized Variance and Volitility
}
