% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heston.R
\name{HestonRealizedVariance}
\alias{HestonRealizedVariance}
\title{Heston Realized Variance}
\usage{
HestonRealizedVariance(realizedVar_0, time, kappa, theta)
}
\arguments{
\item{realizedVar_0}{Initial realized variance of a stock repeated T times}

\item{time}{Number of realized variance observations}

\item{kappa}{Mean reverting parameter.}

\item{theta}{The long term average.}
}
\value{
Realized variance according to the Heston model when the correct kappa and theta values are used
}
\description{
The Heston model's approximation for realized variance.
}
