% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heston.R
\name{HestonRealizedVolitility}
\alias{HestonRealizedVolitility}
\title{Heston Realized Volitility}
\usage{
HestonRealizedVolitility(realizedVar_0, time, kappa, theta, gamma)
}
\arguments{
\item{realizedVar_0}{Initial realized variance of a stock repeated T times}

\item{time}{Number of realized variance observations}

\item{kappa}{Mean reverting parameter.}

\item{theta}{The long term average.}

\item{gamma}{The volatiltiy of the volatility.}
}
\value{
Realized volitility according to the Heston model when the correct kappa, theta, and gamma values are used
}
\description{
The Heston model's approximation for realized volitility.
}
