% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heston.R
\name{HestonVariance}
\alias{HestonVariance}
\title{Heston Variance of Realized Variance}
\usage{
HestonVariance(realizedVar_0, time, kappa, theta, gamma)
}
\arguments{
\item{realizedVar_0}{Initial realized variance of a stock repeated T times}

\item{time}{Number of realized variance observations}

\item{kappa}{Mean reverting parameter.}

\item{theta}{The long term average.}

\item{gamma}{The volatiltiy of the volatility.}
}
\value{
The variance of the realized variance according to the Heston model when correct kappa, theta, and gamma values are used
}
\description{
The variance of the realized variance of a stock according to
the Heston model.
}
