% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HullWhite.R
\name{HullWhite}
\alias{HullWhite}
\title{Hull-White Realized Variance and Volitility}
\usage{
HullWhite(ticker, start_date, end_date, prediction_period, days_to_drop, kappa, zeta, variance, testing)
}
\arguments{
\item{ticker}{Stock ticker}

\item{start_date}{Starting date of data to use}

\item{end_date}{Ending date of data to use}

\item{prediction_period}{How many days to predict using the model}

\item{days_to_drop}{Number of days of realized variances to drop when fitting the model}

\item{kappa}{The drift coefficient of the model.}

\item{zeta}{The Diffusion coefficient of the model.}

\item{variance}{Boolean where true predicts variance and false predicts volitility}

\item{testing}{Whether to use prediction_period days prior to end_date as the prediction period. If false, the prediction period takes place after end_date}
}
\value{
Realized variance or volitility according to the Hull-White model when the correct parameter values are used
}
\description{
The Hull White model is used to estimate the deviation from the long term average of the log return of the stock market
}
