% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HullWhite.R
\name{HullWhiteRealizedVariance}
\alias{HullWhiteRealizedVariance}
\title{Hull-White Realized Variance}
\usage{
HullWhiteRealizedVariance(realizedVar_0, time, kappa)
}
\arguments{
\item{realizedVar_0}{Initial realized variance of a stock repeated T times}

\item{time}{Number of realized variance observations}

\item{kappa}{The drift coefficient of the model.}
}
\value{
Realized variance according to the Hull-White model when the correct kappa value is used
}
\description{
The Hull-White model's approximation for realized variance.
}
