% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HullWhite.R
\name{HullWhiteRealizedVolitility}
\alias{HullWhiteRealizedVolitility}
\title{Hull-White Realized Volitility}
\usage{
HullWhiteRealizedVolitility(realizedVar_0, time, kappa, zeta)
}
\arguments{
\item{realizedVar_0}{Initial realized variance of a stock repeated T times}

\item{time}{Number of realized variance observations}

\item{kappa}{The drift coefficient of the model.}

\item{zeta}{The Diffusion coefficient of the model.}
}
\value{
Realized volitility according to the Hull-White model when the correct kappa and zeta values are used
}
\description{
The Hull-White model's approximation for realized volitility.
}
