% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HullWhite.R
\name{HullWhiteVariance}
\alias{HullWhiteVariance}
\title{Hull-White Variance of Realized Variance}
\usage{
HullWhiteVariance(realizedVar_0, time, kappa, zeta)
}
\arguments{
\item{realizedVar_0}{Initial realized variance of a stock repeated T times}

\item{time}{Number of realized variance observations}

\item{kappa}{The drift coefficient of the model.}

\item{zeta}{The Diffusion coefficient of the model.}
}
\value{
The variance of the realized variance according to the Hull-White model when correct kappa and zeta values are used
}
\description{
The variance of the realized variance of a stock according to
the Hull-White model.
}
